Добавляет деление на модули
